requirements.txt
psycopg2-binary==2.9.9
faker==20.1.0
python-dotenv==1.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
